[gui] move common providers to separate file
